+++ src/components/rootPage/Posters/index.js
@@ -1,3 +1,2 @@
 export { default as Posters } from './Posters';
 export { default as Poster } from './Poster';
-export { default as PosterSkeleton } from './PosterSkeleton';
