+++ src/loaders/index.js
@@ -1,1 +1,1 @@
-// Suspence fallback loader 들 모아주세요
+export { default as PosterSkeleton } from './PosterSkeleton';
